requirements.txt
pandas
requests
mysql-connector-python
sqlalchemy
google-cloud-bigquery
google-auth
streamlit
plotly